+++ apps/astarte_housekeeping/test/astarte_housekeeping_web/controllers/realm_controller_test.exs
@@ -47,6 +47,13 @@ defmodule Astarte.HousekeepingWeb.RealmControllerTest do
   @local_datacenter Repo.one!(from(l in "system.local", select: l.data_center))
 
   @create_attrs %{"data" => %{"realm_name" => "testrealm", "jwt_public_key_pem" => pubkey()}}
+  @create_attrs_with_retention %{
+    "data" => %{
+      "realm_name" => "testrealm",
+      "jwt_public_key_pem" => pubkey(),
+      "datastream_maximum_storage_retention" => 67_000
+    }
+  }
   @explicit_replication_attrs %{
     "data" => %{
       "realm_name" => "testrealm2",
@@ -149,6 +156,53 @@ defmodule Astarte.HousekeepingWeb.RealmControllerTest do
                  "datacenter_replication_factors" => %{@local_datacenter => 1},
                  "device_registration_limit" => nil,
                  "datastream_maximum_storage_retention" => nil
+               }
+             }
+
+      Database.teardown_realm_keyspace(@create_attrs["data"]["realm_name"])
+    end
+
+    test "renders realm when data is valid and retention is set as default for instance", %{
+      auth_conn: conn
+    } do
+      Mimic.stub(Config, :default_datastream_maximum_storage_retention, fn -> {:ok, 30_000} end)
+
+      conn = post(conn, realm_path(conn, :create), @create_attrs)
+      assert response(conn, 201)
+      conn = get(conn, realm_path(conn, :show, @create_attrs["data"]["realm_name"]))
+
+      assert json_response(conn, 200) == %{
+               "data" => %{
+                 "realm_name" => @create_attrs["data"]["realm_name"],
+                 "jwt_public_key_pem" => @create_attrs["data"]["jwt_public_key_pem"],
+                 "replication_class" => "NetworkTopologyStrategy",
+                 "datacenter_replication_factors" => %{@local_datacenter => 1},
+                 "device_registration_limit" => nil,
+                 "datastream_maximum_storage_retention" => 30_000
+               }
+             }
+
+      Database.teardown_realm_keyspace(@create_attrs["data"]["realm_name"])
+    end
+
+    test "renders realm when data is valid and contains retention override, retention is set as default for instance",
+         %{
+           auth_conn: conn
+         } do
+      Mimic.stub(Config, :default_datastream_maximum_storage_retention, fn -> {:ok, 30_000} end)
+
+      conn = post(conn, realm_path(conn, :create), @create_attrs_with_retention)
+      assert response(conn, 201)
+      conn = get(conn, realm_path(conn, :show, @create_attrs["data"]["realm_name"]))
+
+      assert json_response(conn, 200) == %{
+               "data" => %{
+                 "realm_name" => @create_attrs["data"]["realm_name"],
+                 "jwt_public_key_pem" => @create_attrs["data"]["jwt_public_key_pem"],
+                 "replication_class" => "NetworkTopologyStrategy",
+                 "datacenter_replication_factors" => %{@local_datacenter => 1},
+                 "device_registration_limit" => nil,
+                 "datastream_maximum_storage_retention" => 67_000
                }
              }